refactor(jolt-instructions): rename virtual_ module to virt
`virtual` is a reserved keyword; `virt` is the standard abbreviation.
