+++ crates/jolt-instructions/src/virt/arithmetic.rs
