+++ crates/jolt-instructions/src/virt/assert.rs
