+++ crates/jolt-instructions/src/virt/bitwise.rs
