+++ crates/jolt-instructions/src/virt/byte.rs
