+++ crates/jolt-instructions/src/virt/extension.rs
