+++ crates/jolt-instructions/src/virt/xor_rotate.rs
